--- a/docs/README.md
+++ b/docs/README.md
@@ -24,7 +24,7 @@ You can find a complete list of all of our changes and enhancements in [**its ow
 
 Please follow the [**installation guide**](INSTALL.md) to download and install fheroes2.
 
-<a href="https://github.com/ihhub/fheroes2/releases"><img loading="lazy" src="https://img.shields.io/github/downloads/ihhub/fheroes2/total.svg" alt="Github Downloads"></a>
+<a href="https://github.com/ihhub/fheroes2/releases"><img loading="lazy" width="106" height="20" src="https://img.shields.io/github/downloads/ihhub/fheroes2/total.svg" alt="Github Downloads"></a>
 
 ## Copyright
 
@@ -61,18 +61,18 @@ To build the [website](https://ihhub.github.io/fheroes2/) from source, please fo
 We accept donations via [**Patreon**](https://www.patreon.com/fheroes2), [**PayPal**](https://www.paypal.com/paypalme/fheroes2) or [**Boosty**](https://boosty.to/fheroes2).
 All donations will be used only for the future project development as we do not consider this project as a source of income by any means.
 
-<a href="https://www.patreon.com/fheroes2"><img loading="lazy" src="https://img.shields.io/badge/Donate-Patreon-green.svg" alt="Patreon Donate"></a>
-<a href="https://www.paypal.com/paypalme/fheroes2"><img loading="lazy" src="https://img.shields.io/badge/Donate-PayPal-green.svg" alt="Paypal Donate"></a>
-<a href="https://boosty.to/fheroes2"><img loading="lazy" src="https://img.shields.io/badge/Donate-Boosty-green.svg" alt="Boosty Donate"></a>
+<a href="https://www.patreon.com/fheroes2"><img loading="lazy" width="102" height="20" src="https://img.shields.io/badge/Donate-Patreon-green.svg" alt="Patreon Donate"></a>
+<a href="https://www.paypal.com/paypalme/fheroes2"><img loading="lazy" width="96" height="20" src="https://img.shields.io/badge/Donate-PayPal-green.svg" alt="Paypal Donate"></a>
+<a href="https://boosty.to/fheroes2"><img loading="lazy" width="96" height="20" src="https://img.shields.io/badge/Donate-Boosty-green.svg" alt="Boosty Donate"></a>
 
 ## Contacts
 
 Follow us on social networks: [**Facebook**](https://www.facebook.com/groups/fheroes2) or [**VK**](https://vk.com/fheroes2).
 We also have a [**Discord**](https://discord.gg/xF85vbZ) server to discuss the development of the project.
 
-<a href="https://www.facebook.com/groups/fheroes2"><img loading="lazy" src="https://img.shields.io/badge/Facebook-blue.svg" alt="Facebook"></a>
-<a href="https://vk.com/fheroes2"><img loading="lazy" src="https://img.shields.io/badge/VK-blue.svg" alt="VK"></a>
-<a href="https://discord.gg/xF85vbZ"><img loading="lazy" src="https://img.shields.io/discord/733093692860137523.svg?label=&amp;logo=discord&amp;logoColor=ffffff&amp;color=7389D8&amp;labelColor=6A7EC2" alt="Discord"></a>
+<a href="https://www.facebook.com/groups/fheroes2"><img loading="lazy" width="61" height="20" src="https://img.shields.io/badge/Facebook-blue.svg" alt="Facebook"></a>
+<a href="https://vk.com/fheroes2"><img loading="lazy" width="25" height="20" src="https://img.shields.io/badge/VK-blue.svg" alt="VK"></a>
+<a href="https://discord.gg/xF85vbZ"><img loading="lazy" width="93" height="20" src="https://img.shields.io/discord/733093692860137523.svg?label=&amp;logo=discord&amp;logoColor=ffffff&amp;color=7389D8&amp;labelColor=6A7EC2" alt="Discord"></a>
 
 ## Frequently Asked Questions (FAQ)