--- a/docs/_layouts/default.html
+++ b/docs/_layouts/default.html
@@ -39,9 +39,17 @@
     {% endif %}
 
     <link
-      rel="stylesheet"
+      href="{{ '/assets/css/fonts.css?v=' | append: site.github.build_revision | relative_url }}"
+      rel="preload"
+      as="style"
+      onload="this.rel='stylesheet';this.removeAttribute('onload');"
+    />
+
+    <link
       href="{{ '/assets/css/style.css?v=' | append: site.github.build_revision | relative_url }}"
-      media="screen"
+      rel="preload"
+      as="style"
+      onload="this.rel='stylesheet';this.removeAttribute('onload');"
     />
 
     <!-- Scripts -->