--- a/docs/_layouts/default.html
+++ b/docs/_layouts/default.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=640" />
+    <meta name="viewport" content="width=device-width, initial-scale=1">
 
     <meta name="author" content="{{ site.github.owner_name }}" />
     <!-- Open Graph Meta Tags -->