--- a/docs/_sass/_fonts.scss
+++ b/docs/_sass/_fonts.scss
@@ -2,6 +2,7 @@
 
 // Asul font
 @font-face {
+  font-display: swap;
   font-family: "Asul";
   font-style: normal;
   font-weight: 400;
@@ -9,6 +10,7 @@
 }
 
 @font-face {
+  font-display: swap;
   font-family: "Asul";
   font-style: normal;
   font-weight: 700;
@@ -17,6 +19,7 @@
 
 // Lancelot font
 @font-face {
+  font-display: swap;
   font-family: "Lancelot";
   font-style: normal;
   font-weight: 400;
@@ -25,6 +28,7 @@
 
 // Jacquard 24 font
 @font-face {
+  font-display: swap;
   font-family: "Jacquard 24";
   font-style: normal;
   font-weight: 400;