--- a/docs/_sass/_fonts.scss
+++ b/docs/_sass/_fonts.scss
@@ -6,7 +6,7 @@
   font-family: "Asul";
   font-style: normal;
   font-weight: 400;
-  src: url("../fonts/Asul/Asul-Regular.ttf") format("truetype");
+  src: url("../fonts/Asul/Asul-Regular.woff2") format("woff2");
 }
 
 @font-face {
@@ -14,7 +14,7 @@
   font-family: "Asul";
   font-style: normal;
   font-weight: 700;
-  src: url("../fonts/Asul/Asul-Bold.ttf") format("truetype");
+  src: url("../fonts/Asul/Asul-Bold.woff2") format("woff2");
 }
 
 // Lancelot font
@@ -23,7 +23,7 @@
   font-family: "Lancelot";
   font-style: normal;
   font-weight: 400;
-  src: url("../fonts/Lancelot/Lancelot-Regular.ttf") format("truetype");
+  src: url("../fonts/Lancelot/Lancelot-Regular.woff2") format("woff2");
 }
 
 // Jacquard 24 font
@@ -32,5 +32,5 @@
   font-family: "Jacquard 24";
   font-style: normal;
   font-weight: 400;
-  src: url("../fonts/Jacquard_24/Jacquard24-Regular.woff2") format("woff2");
+  src: url("../fonts/Jacquard_24/Jacquard24-Regular-latin-only.woff2") format("woff2");
 }