--- a/docs/_sass/_fonts.scss
+++ b/docs/_sass/_fonts.scss
@@ -32,5 +32,6 @@
   font-family: "Jacquard 24";
   font-style: normal;
   font-weight: 400;
-  src: url("../fonts/Jacquard_24/Jacquard24-Regular-latin-only.woff2") format("woff2");
+  src: url("../fonts/Jacquard_24/Jacquard24-Regular-latin-only.woff2")
+    format("woff2");
 }