--- a/docs/assets/css/style.scss
+++ b/docs/assets/css/style.scss
@@ -1,7 +1,6 @@
 ---
 ---
 
-@import "fonts";
 @import "dark-theme";
 @import "light-theme";
 @import "layout";