## Overview
This PR introduces several optimizations to improve website performance, accessibility, and user experience. The changes focus on modernizing the website's assets and implementing best practices for web development.

## Key Changes

### Performance Optimizations
- Converted all font files from TTF to WOFF2 format for better compression and faster loading
- Added `font-display: swap` to all font declarations to improve text rendering performance
- Converted PNG screenshots to WebP format for better compression while maintaining quality
- Added `loading="lazy"` to images for better page load performance
- Implemented proper viewport meta tag for better mobile responsiveness
- Added `defer` attribute to JavaScript files to improve page load time
- Separated font CSS into its own file and implemented preload for critical CSS

### Accessibility Improvements
- Added proper alt text to badges
- Improved image aspect ratios for better visual consistency

### UI/UX Improvements
- Adjusted font sizes for better readability
- Improved responsive layout handling
- Enhanced image showcase layout with proper aspect ratios

# Performance Comparison

My total load times are regularly approaching 220ms, with no cache. However, I also used webpagetest.org to run repeatable benchmarks with timing breakdowns.

## 3G connection in Mumbai, IN

### Web test links
[Current performance](https://www.webpagetest.org/result/250511_AiDcVK_63K/)
[New performance](https://www.webpagetest.org/result/250511_AiDcX9_641/)

```
Connectivity: 1600/768 Kbps, 300ms Latency
```
### Results

| Test | Current | New |
|--------|--------|--------|
| LCP | 2.623s | 2.152s |
| First view | 6.057s | 5.110s |
| CLS | 1.047s | 1.062s | 
| Page Weight | 646KB | 562KB |

## Cable load times

### Web test links
[Current performance](https://www.webpagetest.org/result/250511_BiDcAP_5R3/)
[New performance](https://www.webpagetest.org/result/250511_AiDcGB_5YQ/)

```
Connectivity: 5000/1000 Kbps, 28ms Latency
```

### Results

| Test | Current | New |
|--------|--------|--------|
| LCP | 1.13s | 0.296s |
| First view | 1.507s | 0.333s |
| CLS | 0.004s | 1.026s | 
| Page Weight | 646KB | 562KB |

## Native load times

### Web test links
[Current performance](https://www.webpagetest.org/result/250511_AiDcVQ_5ZA/)
[New performance](https://www.webpagetest.org/result/250511_BiDc6R_5Z8/)

```
Connectivity: Native
```
### Results

| Test | Current | New |
|--------|--------|--------|
| LCP | 0.474s | 0.286s |
| First view | 0.431s | 0.264s |
| CLS | 0s | 1.036s | 
| Page Weight | 646KB | 562KB |

# Future performance benefits

These are performance benefits that are still available, which haven't been implemented with this work. Mostly because they would require a build step.

1. Inline critical styles into the HTML to reduce CLS (cumulative layout shift) in initial page load (via tool like critical) and minify the rest
2. Minify Javascripts using build process (via tool like terser)
3. Implement a real CDN in front of assets with a longer TTL for static assets (Github pages has a SHORT TTL and this makes the CDN less effective).