@oleg-derevenetz 2025-05-11
> Implement a real CDN in front of assets with a longer TTL for static assets (Github pages has a SHORT TTL and this makes the CDN less effective).

As far as I understand, this will require using another hosting instead of GitHub Pages, because GitHub Pages does not allow to manipulate HTTP headers, right? I suppose we would like to avoid this.