@justinbburris 2025-05-11
> > Implement a real CDN in front of assets with a longer TTL for static assets (Github pages has a SHORT TTL and this makes the CDN less effective).
> 
> As far as I understand, this will require using another hosting instead of GitHub Pages, because GitHub Pages does not allow to manipulate HTTP headers, right? I suppose we would like to avoid this.

Yep - let me expand on this point

1. Github pages don't have great TTLs for their headers. This is _mostly_ fine but it's unfortunate in that we'll get constant cache misses when they time out. So it's fine for a session, but not great for multiple sessions. This may not be a concern (and considering we're already doing great with our performance, it probably is)
2. If it was a big enough concern, we have 2 paths forward
  a. Serve via another hosting provider
  b. Use a domain to proxy to GH pages

Both options are likely non-starters, but I figured it'd be good to call it out as another optimization that could be done in the future.