@justinbburris 2025-05-12
> Hi @justinbburris technically looks good to me, except for one thing: you need to synchronize changes between `docs/README.md` and `README.md`. Otherwise, the new webp images will not be displayed in the repository description on GitHub itself and the wording will be different.

Gah! I had that fixed in a commit I dropped (to clear out some other unrelated changes) 

Sorry for the sloppy submission. I'll get that fixed. 

🙇