@justinbburris 2025-05-13
> Hi @justinbburris , I have unrelated to this pull request question: what do you think about generation of web pages from wiki section of this project? Users shouldn't be "redirected " to dev related pages when they browse the project's website.

Agreed - and this is tricky. The wiki feature, afaik, cannot be directly embedded into the Github Pages site as is. What we could look into is one of three paths. I've ordered this list from what I'd recommend, given the tradeoffs.

### Option 1. Completely move away from Github Wikis and move into docs in repo (and thus natively into Github Pages)

This _could_ be generally good as the wiki is locked to changes to a subset of users. Putting the content in the repo, the number of contributors would naturally rise, without going through a central source. Some pages could be converted to other Github features, like projects, etc.

### Option 2. Port specific pages into Github pages

If the page is linked form the doc site, then they should live in the docs

### Option 3. Use a tool to convert our wiki into pages embedded into the doc site at build time.

This would require a build step (we're talking about doing this anyway) and would be a bit tricky to implement in such a way that the devex makes sense, but we could go this path.