@justinbburris 2025-05-13
Oh Good catch on the alt-text. Nice to make that improvement everywhere.